ROX-35685: move mapping validate/hash to pkg/virtualmachine/cpemapping
pkg/scannerv4/repositorytocpe is scanner-team territory; the validation
and hashing helpers are specific to the VM-scanning VSOCK sync protocol,
not a general scanner v4 concern. Move them to a VM-domain package that
both roxagent and sensor can import without crossing ownership boundaries.

AI-assisted.
