+++ pkg/virtualmachine/cpemapping/validate_test.go
@@ -1,4 +1,4 @@
-package repositorytocpe
+package cpemapping
 
 import (
 	"bytes"
